chore: fix float16 accumulation, drop unused peft, enable full clips on 48GB GPU
- strategy_maml.py: keep θ* in float32 throughout; the float16 cast after each
  round was accumulating quantization error over 200 rounds with no memory benefit
  on a 48GB machine
- requirements.txt: remove peft pin (~300MB install); LoRAWav2Vec2 uses manual
  LoRA because peft.LoraModel is incompatible with higher.innerloop_ctx
- vctk_lora_poc.yaml: set max_audio_samples=null and support_size=20 for the
  48GB GPU (full 7s clips, larger support set)

+++ poc/federated/strategy_maml.py
@@ -102,10 +102,8 @@ class PerFedAvgStrategy(fl.server.strategy.Strategy):
             gc.collect()
 
         for i in range(n_params):
-            avg = acc_grads[i].astype(np.float32) / total_weight
-            self._theta_star[i] = (
-                self._theta_star[i].astype(np.float32) - self.outer_lr * avg
-            ).astype(np.float16)
+            avg = acc_grads[i] / total_weight
+            self._theta_star[i] = self._theta_star[i].astype(np.float32) - self.outer_lr * avg
 
         del acc_grads
         gc.collect()
